mip_generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mip_generator"
version = "0.1.0"
description = "A project to reproduce the research from 'Attacking Multimodal OS Agents with Malicious Image Patches'."
requires-python = ">=3.10"
dependencies = [
    "torch",
    "transformers",
    "pillow",
    "accelerate",
    "tqdm",
    "wandb",
]
